refactor: reformat the package following industry convention
