+++ src/main/java/betterbanking/controller/TransactionController.java
@@ -1,6 +1,7 @@
-package betterbanking;
+package betterbanking.controller;
 
-import org.springframework.stereotype.Controller;
+import betterbanking.entity.Transaction;
+import betterbanking.service.TransactionService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
