+++ src/main/java/betterbanking/entity/Transaction.java
@@ -1,4 +1,4 @@
-package betterbanking;
+package betterbanking.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
