+++ src/test/java/betterbanking/TransactionComponentTest.java
@@ -1,9 +1,12 @@
 package betterbanking;
 
+import betterbanking.controller.TransactionController;
+import betterbanking.entity.Transaction;
 import betterbanking.model.*;
 import betterbanking.repository.InMemoryMerchantDetailsRepository;
 import betterbanking.repository.RESTTransactionsAPIClient;
 import betterbanking.repository.TransactionApiClient;
+import betterbanking.service.TransactionService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import okhttp3.mockwebserver.MockResponse;
